arm: dts: Remove incorrect /memory node in skeleton
This node is not correct (address of 0, size of 0) but makes it look like
we have a valid memory node. Remove it. Boards should add a correct memory
node if they want one.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot an image using the latest cros_bundle_firmware, with
early-firmware-selection support (although not enabled using ToT U-Boot).
See that it writes to the correct memory address on the board, and it boots
correctly to the kernel.

Without this change, cros_bundle_firmware things SDRAM starts and 0 and
writes to the wrong address.

Change-Id: Id82df9fc719a626d95471a778692d72026d42680
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64866
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>