elog: Add configuration information to the device tree.
The full size of the log, the size to shrink it by when it's full, and a few
other parameters are stored in the device tree to be retrieved by the event
log code at run time.

BUG=chrome-os-partner:16132
TEST=With this and other changes, booted on pit and used mosys eventlog
list to look at the events in the log.
BRANCH=pit

Change-Id: Iffa0350b10f009c98eddb5c78d277acc3f878fdb
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/65166
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>