cros: Adjust cros_fdtdec_memory() to allow multiple memory banks
The current code assumes there is only one memory bank, but this is not
always true. Adjust it to permit multiple banks, even though it will
only return the first.

This function will be deprecated when the early-firmware-selection
feature is in.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Boot on Pit with the memory bank patches, in particular:

https://gerrit.chromium.org/gerrit/#/c/64106/

See that verified boot completes as before.

Change-Id: I4e8b2e3fa0369808d82b77787fcd9fe4a9ab102c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64971
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>