efs: arm: Allow SPL to self-relocate
Allow SPL to run from any location, by adding code to copy itself on start-up
if it finds the location is wrong. It does this by comparing its expected
start with the actual start calculated from the current PC value.

This is used for the resume path of early-firmware-selection. We run the RW
SPL at a different location in IRAM, the place we stash it to on boot. That
location is near the top of IRAM.

In the resume path, we run BL1, RO SPL, then RW SPL. The RW SPL sees that it
is in the wrong place, copies itself down over the top of RO SPL and then
continues to run.

This is easier than using a relocatable binary, smaller adding relocation code
to SPL, and ensures that SPL always runs at the same address - easier to
understand and debug.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
With other patches, cause the RW SPL to be executed, and see that it runs
correctly. Without this code it dies in board_init_f() when it loads from the
literal pool.

Change-Id: Ie1b2eebf382f2203da3976bc908cd30a455815c4
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65193