efs: exynos: Add an accessible header for the ACE SHA engine
The hashing function currently has not API accessible from the rest of
U-Boot. Add this, so we can make use of this acceleration engine.

Also make it buildable from SPL with a suitable option. This will remove
the need to add hashing code to SPL.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
With other changes, see that RW SPL verifies U-Boot after loading, and
prints an error if the hash is wrong (test this by manually changing
the hash in the SPL with ghex).

Change-Id: If90a8de67a428b4a6a511b676c9b4d137dd1ba75
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65194