efs: arm: Allow cache to be enabled for regions < 1MB
On-chip SRAM is often less than 1MB in size, but it might be quite slow to
access. To get around this, we want to enable the caches for this region
of memory also.

Adjust the code so that it enables the cache for regions less than 1MB.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot RO U-Boot from IRAM on Pit. See that it goes through
verified boot much more quickly now. When it boots to RW U-Boot, press
a key to stop. Use the 'md' command to check the MMU table and see that
there is a cached entry for the IRAM region.

Change-Id: I22b9147cdd30155f3d6f8e73c554dc6be2782767
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65195