cros: Simplify the flashmap data structures and code
We need to add a few more images to the flashmap, so before doing so we
should sort out the horrible code.

The current data structure has a different structure for read-only and
read-write. The differences are not that great, and it is much simpler
to use the same structure.

The hierarchical multi-blob images we used on snow are no-longer needed.
We no longer use them for the EC (which has its own section now) and we
don't want to start using them for SPL (which will need its own section).
A recent change to cros_bundle_firmware ensures that hashes are added
to the node they refer to, no just the sub-nodes.

The FDT is still currently appended to U-Boot so this is a multi-blob
image, but nothing needs to access it since it is loaded with U-Boot.
If and when we switch to a FIT image for U-Boot (loaded by SPL) then
everything will be in its own image.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual

Change-Id: I13fe5c58aebb17677215a5b9488d4d616fce1d78
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65197