cros: Add new fmap regions for SPL and recovery
With early-firmware-selection we need to have a separate SPL for each RW
image, since it is possible that SDRAM will not be set up when these are
started.

The recovery image needs to be separate from the RO U-Boot, since the RO
U-Boot is a cut-down image unable to use the LCD, USB, etc. It therefore
needs its own section in the image.

Add this information to the flashmap along with associated decode and
test code.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual

Change-Id: I463278e2120bdc91322d7b6dc60d9dfa7e064f4c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65198