efs: exynos: Add parameters for using a second SPL
In some situations it is convenient to be able to jump from the main SPL
to a secondary SPL. One example of this is when the main SPL is in
read-only memory and the secondary SPL is in read-write memory. In this case
it becomes possible to upgrade the code that does SDRAM init in the secondary
SPL. In fact the main SPL could avoid initializing SDRAM altogether.

With suspend/resume we also want to be able to use the RW SPL. To achieve
this, it can be stashed into a location in IRAM, and then executed by the
RO SPL in the resume path. This means that memory resume code can also be
made updatable.

Add machine parameters to indicate the location of the secondary SPL in
IRAM, its size and whether it should be used by RO SPL. Also add a parameter
to indicate whether a particular SPL should initialize DRAM or not.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
With following CLs, see that the RW SPL is used in a recovery or RW boot,
and that it is also used in the resume path (by corrupting it in IRAM and
seeing that s2c then fails).

Change-Id: I955ecb9bed031f243e17d3bf3bb705bab1ec182e
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65199