efs: exynos: Refactor SPL to make wakeup configurable
At present the decision about what low-level init to do is done in the
same place as the init itself. This means that it is not possible to find
out what is intended to be done, and adjust it.

Split the two parts into separate functions and call a board-specific wakeup
processing function on wakeup (this will be used to jump to RW SPL if
early-firmware-selection is enabled.)

In the case of a wakeup we don't need to run all the clock init again, so
change this also.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on pit with RW firmware, see that everything looks normal.
Run suspend_stress_test to check that resume still operates as expected.

Change-Id: Ic293c4b9d5a127432b025fc549b64f43ea86a0be
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65202