efs: exynos: Round up to nearest block size when reading U-Boot
For boot media with a block size (such as mmc) we must round up the number
of bytes we read so that we read the entire last block. Add code to do
this for eMMC and SD card.

Previously this was done in cros_bundle_firmware, but it means that SPL
does not have an accurate size for U-Boot, which makes hashing it for
verification difficult.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual

Change-Id: I33dc1958a25cdb36d958eafcf681e11e87f1d12e
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65200