split: efs: cros: exynos: Use a parameter to indicate RO/RW firmware
The current code uses a word (written to the initial stack by SPL) to
indicate that U-Boot is the RO firmware. This word is overwritten in
U-Boot so that the subsequent RW U-Boot knows that it is RW firmware.

This approach does not work when we have two SPLs in the boot chain.
With early-firmware-selection we have two SPLs, one of which needs to
indicate to its U-Boot that it is RO firmware, and one of which needs
to indicate to its U-Boot that it is RW firmware.

A more robust approach is to just pass a parameter to U-Boot. This is
supported by the U-Boot start-up code.

A special value (not used by BL1) is used to indicate that we are running
from another U-Boot rather than SPL. This is an indication of RW firmware.
When U-Boot runs SPL, the parameter is picked up by SPL and passed on to
the U-Boot that it loads.

This method also works without early-firmware-selection, where there is
no RW SPL.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on Pit with and without EFS using --force-rw. See that
in both cases vboot goes through the correct stages and boots.

Change-Id: Ibaa36767dccf0bb8b67e3bb01673b99ecdf7574f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65201