efs: gen: Adjust the macro for removing commands from U-Boot
Some subcommands can have the same name as others. This works because they
are hidden behind a top-level command. But this causes problems with the
method used to remove commands from U-Boot when CONFIG_CMDLINE is not
defined.

Adjust the macro slightly so that duplicate subcommands are supported.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
See no more build failure with CROS_RO defined.

Change-Id: I065238df621855cbda9db4d0d3a3ba9dd4bc52ba
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65203