efs: gen: Use correct start symbol for U-Boot
CONFIG_SYS_MONITOR_BASE may not be defined for all boards. A better choice
for the start of the U-Boot code is _start which is defined in the sections.h
header file and should be available on all boards.

On sandbox, _start is available but incorrect. The sandbox code is not
actually stored in the U-Boot memory map, but leave the code as is for now.

BUG=none
BRANCH=pit
TEST=manual
Build and boot on Pit and sandbox. Build for link.

Change-Id: I191f54746a1527538b4a849c733690f23bef08fb
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65483