efs: exynos: Allow verification after loading U-Boot
Rather than just blindly running the U-Boot that SPL loads, we can check it
against a stored hash to verifiy it. The hash is attached to the SPL by
the build system and can be accessed by a function provided for the
purpose.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
With other changes...

Boot with EFS on pit with '#define DEBUG' at the top of spl_boot.c and see
that it prints a verify message on RW boot.

Manually corrupt the SPL hash (last few bytes of the SPL image) and see
that it now fails to verify.

Change-Id: I413f370cb62f76808bc20d9bda1cbaae336e8bdf
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65485
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>