efs: cros: Introduce cros_fdtdec_decode_region() to find a memory region
The existing cros_fdtdec_alloc_region() has an allocation feature which we
really don't need. It was only used for sandbox, and sandbox only needed it
because it didn't have a proper memory model.

Introduce a new function which does not allocate, but allows the region
to be in a memory region other than /memory. It also handles appending the
"-offset" string to the property name, as well a suffix if required.

This feature can be used like this:

chromeos-config {
	google-binary-block-offset = <0x01200000 0x100000>;
};

This means that the GBB will be placed at offset 0x1200000 into SDRAM, as
specified by the /memory node:

memory {
	reg = <0x20000000 0x20000000>;
};

So in this case the resulting address will be 21200000, and the size of the
region will be 1MB,

It is also possible to specify the memory region (here "/memory") and a
suffix to add, so that the property can be modified, like:

chromeos-config {
	google-binary-block-offset = <0x01200000 0x100000>;

	google-binary-block-memory,efs = "/iram";
	google-binary-block-offset,efs = <0x02200000 0x100000>;

	google-binary-block-memory,backup = "/highram";
	google-binary-block-offset,backup = <0x03200000 0x100000>;
};

assuming suitable definitions for /iram and /highram (containing a 'reg'
property just like a standard /memory node).

This function is backwards compatible with the old function except for the
unused allocation feature, but provides flexibility required by
early-firmware-selection.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on pit - see that GBB, crossytem and the kernel are all
loaded correctly and the system boots through RW to a kernel.

With EFS verify that the crossystem FDT region is correct located.

Change-Id: I9bf72ac988df76770933cdc589740d52306bf46e
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65486