efs: cros: Adjust nvstorage to use a name instead of a number
There is no good reason to use a number here. It makes it harder to add
new types (when RO firmware requires an old type) and is not in keeping with
the way devices tree works. A string is almost as easy to store.

In preparation for moving to device tree for crossystem data, change this
field to a string and clean up the code a little.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Boot on pit and see that nvcontext still works correctly - mosys works,
recovery can be selected using crossystem in the kernel

$ mosys nvram vboot read
7000000000002600000000000000003c

$ crossystem recovery_request=5

(reboot and it goes to recovery)

Change-Id: I8b488e91d3f3e461354a383e77b17f3167a3a4a0
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65487