efs: cros: Display a message on each boot_device read
Display a message when data is read from the boot_device.

The code for boot_device is not quite correct - in the event of an error
the timing is wrong. This is a very minor bug, but we may as well fix it
so this bad example is not in the code. Ensure that bootstage time does
not include any text output delay.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on pit, see that the bootstage report shows the time spent
reading from MMC.

Change-Id: I46ced0be4e485c37100aa9ad6687d6a7e4316506
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65489