efs: cros: Use correct API to access host device for sandbox
There is a new function for obtaining the host device which does not print
errors to the console. Since vboot regularly asks for devices which do not
exist (as part of its scan for available disks) this helps to reduce console
spam. Adjust the code to use the correct call.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
$ crosfs -b sandbox -V
$ rm x.dtb; b/chromeos_sandbox/u-boot -d b/chromeos_sandbox/out/updated.dtb \
	-c "sb bind 0 /home/sjg/c/src/build/images/peach_pit/latest/chromiumos_image.bin; vboot_twostop"
	--spi_sf 0:0:W25Q128:b/chromeos_sandbox/out/image.bin  -s x.dtb -w \
	-r  -n  -i

Hold down 'Escape' when it starts to enter recovery.

See that we no longer get lots of messages about block devices not being
bound. We care not.

Change-Id: Iec94aef414e23ddd726c9761b186378f045680d8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65490