efs: cros: Enable display for sandbox
Now that sandbox has an emulated LCD display, enable it for verfied
boot.

For some reason the character data in crossystem is unsigned. We can avoid
casts by making it plain 'char'.

BUG=chrome-os-partner:21115
BUG=chromium:244019
BRANCH=pit
TEST=manual
Run sandbox U-Boot and enter recovery mode, see that it displays the recovery
screen with correct colors, etc.

Boot through RW on pit into the kernel and verify the the dev screen
appears (in dev mode).

Change-Id: Ifd59c1b45e5ad796df7c2a38857bbcf9bc0f7e5e
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65491