efs: cros: Add stage-based verified boot
The code in vboot_twostop uses a byzantine initialization path which is hard
to adjust without much use of #ifdef. It also uses a C structure to pass
crossytem data from one U-Boot to the next instead of a device tree, as used
by the kernel, on ARM at least. The code has become quite complex and hard to
understand over time.

With early-firmware-selection we want to make even more changes to this code.
The result would certainly add considerably to the pain. It is time to pull
this code out into separate chunks for each stage, and allow each stage to
stand alone. In particular, we need to avoid compiling in the RW code in the
smaller RO U-Boot.

The new files are as follows:

   cmd_vboot.c - the new cleaned-up verified boot command
   stages.c    - keeps track of the available vboot 'stages'
   readonly.c  - holds functions for the read-only stages
   readwrite.c - holds functions for the read-write stages
   vboot.c     - vboot helper functions for RO and RW
   vboot_fdt.c - reads and writes vboot stage to an FDT
   vboot_dump.c- writes out information about vboot state
   verify.c    - verifies a U-Boot image (from SPL)
   wakeup.c    - handles wakeup through RW SPL

A new 'vboot' command makes these stages avaialble on the command line.

The code largely remains the same, apart from its ability to deal with
early-firmware-selection, but it is much easier to deal with in the new form.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Verified normal, dev mode, recovery mode, switch and other functions manually
tested on Pit.

Change-Id: I8364f72f024490646293bae03fe5b753a60a3193
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65493