efs: exynos: Remove hard-coded addresses from lowlevel.S
The hard-coded address currently match with the config file, but if we
change one and not the other, it will break. Use the same symbol in each
case.

Also remove setting of the stack pointer since the new ARM SPL code does
this automatically. At present the effect is the same but it is best
avoided.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual

Change-Id: I8ab8d21274f13941dbaa48a455171a20a0ca58ce
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65522