arm: Remove unnecessary stack allocation
The new ARM relocation code deals with global_data automatically. We don't
need to create another one.

Prior to relocation U-Boot uses an initial stack which is on IRAM on the
exynos. We are short on initial stack space (0x1c bytes free at present).
Let's see if this change allows us to avoid increasing it for now.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on Pit with (without EFS) and see no ill effects. The stack
overflow checking would print a message if there were any problem with the
stack. Everything boots normally with the standard ARM global_data.

Change-Id: I086860838e3acac93de40906b87a5cf70b9148f9
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65521