efs: exynos: Add a machine param for the vboot persist address
SPL needs to know the location of this memory. Add a machine_param for it
which is filled in by cros_bundle_firmware.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Using 'cros_bundle_firmware -v3' to see that the correct address is picked
up.

Using SPL debug code, see that RW SPL can access this and prints a message
displaying the count for the selected firmware.

Change-Id: Id1c957098aa7c4a070a1b2dd74f427327df13ca7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65537