efs: Allow power_management to be used in SPL
Console recording would be a useful feature in SPL, but for now we don't have
the instructure to pass the recording to U-Boot, so let's disable it.

Also remove the VBDEBUG() calls since SPL is not linked with the vboot
library.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on pit, see that SPL still runs, and that the SPL

Change-Id: I794991820212d7623e3c84bb6129a1ffaf83a6c7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65538
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>