efs: Avoid using CONFIG_FINAL_LINK
This does not work properly at present - it drops linker lists when it is
enabled, meaning that they do not appear in the image. The option was added
as a way to drop unused code in U-Boot (specifically commands) but it is not
actually required.

This problem was discovered when EFS was completed. During testing the
nvcontext part of vboot, the nvcontext drivers did not get included in the
image so this could not work.

Let's disable it and revisit later if it proves to show a benefit.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and run on Pit - see that the NV context can now be read from the EC.
Previously this driver would be omitted from the image.

Change-Id: Ic41b0b7e9569ce4df4d9940f675887a45fe8de83
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/65905
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>