efs: cros: dts: Add device tree changes for Early Firmware Selection
This change adjusts the flashmap to incorporate a separate recovery U-Boot
as needed by EFS. It also adds a separate SPL for each of the U-Boots (RW-A,
RW-B and Recovery). These are only used when EFS is enabled.

Memory areas are selected for use by EFS also. Note EFS is only supported on
Peach at present (not snow/spring) and additional device tree and config
changes would be needed to enable it for those platforms.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and test normal and EFS firmware on Pit. Run FAFT.

Change-Id: I21d5d417bebcccb2a103a3b1395eebd402751da3
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65548