exynos: limit memory used for passing data to kernel
https://gerrit.chromium.org/gerrit/#/c/65483/4 corrected the size
calculation for U-Boot.  Prior to this, U-Boot reserved 1.1GB of space
for itself on the stack, forcing all other allocations down in the
address space.

With this fixed, and U-Boot only reserving ~700K of space, allocations,
such as the device tree passed to the kernel, were near the top of
memory, outside of lowmem, where the kernel requires the device tree to
be placed.  Because of this, the kernel was unable to properly read the
device tree, and failed to boot on Spring.

http://permalink.gmane.org/gmane.linux.ports.arm.kernel/258992
recommends placing the device tree just beyond the 128MB boundary at the
beginning of memory, which the kernel decompressor cannot exceed.  Thus,
limit the data passed to the kernel to the first 256MB of memory.  This
memory is allocated starting at the top, leaving 128MB of space for
allocations, but staying close to the 128MB boundary.

BUG=chromium:271558
BRANCH=none
TEST=Spring boots kernel successfully

Change-Id: Ib539739d0d7de8e9b8ecf97571f923c2e5209a95
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65745
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>