exynos: initialize global MCT counter at startup
Linux kernel finds out firmware boot time by reading the value of the
global MCT counter. This counter has to be started as early as
possible. This change adds initialization code to set timer counter to
zero and start it.

This is a rebase of ddb8818, there has been a conflict in

arch/arm/cpu/armv7/exynos/lowlevel_init.c,

it turned out to be easier to modify this file "by hand".

BRANCH=none
BUG=chrome-os-partner:21226
TEST=manual
  . with this change compiled in the following can be found in the
    dmesg log:

   localhost ~ # dmesg | grep 'Initial '
   [    0.000000] Initial usec timer 1205824

  before this change the value of the usec timer was 0

Change-Id: I5d8f98a81e61e2c6b4c7aae733a1067a4491c897
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66683
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186133