cros: dts: Select a revision specific device tree from the kernel
If the kernel happens to have a revision specific device tree, we
should pick it.  This is the clean ToT version of the hack at
<http://crosreview.com/167483>.

Note that the current ToT U-Boot doesn't seem to reliably boot the
kernel all the way to the command line anyway on Snow, but that's not
related to this change.

BRANCH=none
BUG=None
TEST=On rev3, see that rev3 device tree is picked up
TEST=On rev3, hack board_get_full_revision() to return +1 and see
that rev4 device tree is picked up.

Change-Id: I9baf15cdd8653524c2a8f876b066263b0453e286
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/167498
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>