gen: usb: xhci: Handle short transfers in multi-TRB TDs correctly
When XHCI controllers encounter a short packet in a TRB that is not the
last TRB of the whole transfer, they generate two transfer events: a
Short Packet for the current TRB and a Success for the last TRB. (Linux
considers this a quirk and calls it XHCI_SPURIOUS_SUCCESS, but I think
it's debatable whether it's really a quirk when the XHCI specification
specifically documents this behavior as expected at the end of section
4.10.1.1.) The U-Boot stack currently does not expect the second event,
which desyncs its perception of the event ring and can lead to all
kinds of nasty BUGs much later on.

This patch changes the transfer cleanup to check for this condition and
look for a second event, which can be combined with an earlier fix for a
similar problem that was specific to control transfers. It will also
scan the remaining list of unprocessed TRBs to figure out the actual
amount of untransferred bytes and return the correct length value, since
the length field of the event only contains untransferred bytes from the
current TRB. Finally, it adds another BUG_ON() that will help detect
desynchronized event rings right away in the future.

BRANCH=pit
BUG=chrome-os-partner:22545
TEST=Make sure netboot, USB boot and recovery still work on Pit.

Change-Id: I93ddd390ed69fec6bead1b72fd9bf6f2b37cd2b4
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168520