gen: mmc: Use correct divisor when requested freq > src clock
The current code assums that the source clock is at least as large as
the requested MMC frequency, so always introduces a minimum divisor
of 2. However this is not always the case.

Deal with this as a special case.

BUG=chrome-os-partner:21741
BRANCH=pit
TEST=eMMC and SD boot still works. boot_device_read time decreases
from 189ms to 95ms with RW boot. This is approximately the expected time
(same as Snow). The boot_device_info time also drops.

New RW boot time is:

    858,499          1  start_kernel

Accumulated time:
                 7,026  SPI read
                38,998  boot_device_info
                94,737  boot_device_read

Change-Id: I1f584643fc2a5ba3da955f2003b0af3ad3a1198e
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169123