cros: pit: extend maximum EC wait time to 1.1 second
It turns out that EC behavior has been changed recently
(https://gerrit.chromium.org/gerrit/66703) to address a Linux driver
issue (inability to lock SPI interface). This broke the U-Boot EC
driver, which relied on the ability to receive the 'in progress' EC
state updates.

A simple albeit not the most efficient solution is to extend the
maximum EC transaction time to 1.1 second which is enough for the
worst case scenario erasing of 64K EC flash section.

BRANCH=none
BUG=chrome-os-partner:22368
TEST=manual
   . attempted firmware_SoftwareSync five times, it succeeded

Change-Id: I7d48c338b00985133cf5d5953ad20fc6df1d63ed
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170180
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186477
Reviewed-by: Doug Anderson <dianders@chromium.org>