coreboot: set default baud rate to 115200
The proper way to configure the UART is to inherit the coreboot
settings, and in case coreboot did not enable the console - use the
device tree. This is a quick fix to get the console output going so
that u-boot is usable.

BRANCH=none
BUG=none
TEST=manual
   . with other changes the Bayley Bay board comes up all the way to
     u-boot prompt.

Change-Id: Ie0e0d61c39eb60c413752495527f4fa6ecfde856
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/171575
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>