tegra: venice 1 & 2: spi: fixed CS active bug during reset
This is to fix the CS active bug when the SPI controller is reset.
A new DT property, 'nvidia,cs-pinmux' is needed for the SPI driver
to control the CS line so the line won't be driven out during reset.

Also, move the calling of clock_start_periph_pll() back inside
tegra114_spi_claim_bus().

BUG=nvbug 1372781
BRANCH=none
TEST=test on Venice & Venice 2:
     - do 'crosec init' on u-boot prompt, and note that EC console won't
       display '[SPI rx bad data] and [SPI not ready] error messages;
     - do SPI ROM access (like, 'sf probe 1:0' on u-boot prompt), and
       check that there is no glitches on CS line.

Change-Id: Ide177d9957c051a3bfc42ad7f143d7c90f4735d9
Signed-off-by: Yen Lin <yelin@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/176049
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>