Tegra: MMC: Flip polarity of SD-card CD line on Norrin FFD ref boards
Using the Acer (instead of the Pixel) SD-card adapter means that
the card detect (CD) line is reversed in polarity. Fix it by
using board_get_revision() to determine if we're on a Norrin FFD
(Google reference board), and flip the polarity. Otherwise, we're
on a Venice2 (or other Tegra board0, so we leave it alone.

Untested on Norrin, but it still works as before on Venice2.

BUG=none
TEST=as above

Change-Id: I8ed9b142377f0032da8f8f02c161cb7f10dce70c
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/176402
Reviewed-by: Yen Lin <yelin@nvidia.com>