cros: Add peach-pi; remove ancient peach-pit-rev1 support
pi is a variant of peach (like pit), so start by just making it a copy
of pit with a new compatible string.  Promote the old "peach-pit.dtsi"
file to just a "peach.dtsi" file.  We'll move things out of the
"peach.dtsi" file as needed.  Since we append the revision
automatically now in U-Boot we don't need all of the extra revision
dts files.

While we're at it we'll remove the old "rev1" device tree file.

BRANCH=pit
BUG=chrome-os-partner:24492
TEST=Builds

Change-Id: I8176820112cb493f10562b54ae5feb789d75c434
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179458
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>