x86: Add cleanup_before_linux_select() stub
The cros code is going to use this function in the upcoming patch, and
needs to be able to compile for x86 targets just as well as for arm.
Instead of adding a conditional compilation in common code let's add a
placeholder for the new function in x86 space, we might enhance it in
the future as well, if and when needed.

BRANCH=None
BUG=chrome-os-partner:21115
TEST=none

Change-Id: I1b32693e9cc305ddf261843e50ae0fa2f5467c64
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188138
Reviewed-by: Simon Glass <sjg@chromium.org>