gen: Initialize serial console first
The side effect of not defining CONFIG_SYS_CONSOLE_IS_IN_ENV is
forcing u-boot to use only the first of the registered console
interfaces. This causes disappearing of the serial console output in
case the LCD console is enabled.

Moving serial console initialization to the first position makes sure
that the serial console output does not stop after console driver is
initialized.

BRANCH=None
BUG=chrome-os-partner:25573

TEST=manual
   . the u-boot built with CONFIG_CROS_FULL undefined (which removes
     the CONFIG_SYS_CONSOLE_IS_IN_ENV definition) keeps printing on
     the serial console after console driver is initialized.

Change-Id: I972c1544fabaf693395896bffc722e2a9476c0fa
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/189401
Reviewed-by: Simon Glass <sjg@chromium.org>