exynos: spl: Drop unnecessary SPL action
In the existing code DO_CLOCKS and DO_MEM_INIT SPL actions are
strictly used together, one of the two can be safely eliminated.

The discrepancy could happen if EFS is enabled and it is decided that
the first SPL (started by ROM) should not initialize memory, this
issue could be addressed separately if and when required.

BRANCH=none
BUG=none
TEST=manual
   . pit pi and daisy compile successfully.

Change-Id: Ieb522b9dfabe2dbbbaacffa43a12bbf8f19ccce9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191354
Reviewed-by: Doug Anderson <dianders@chromium.org>