vboot: change VbExEc implementations to support devidx parameter
This will be used in subsequent CLs to support PD software sync.  For
now, only devidx=0 is used, and other values return error.  This makes
static analysis of the change easy.

This must be checked in at the same time as changes to the
vboot_reference project and depthcharge implementation.

BUG=chrome-os-partner:30079
BRANCH=none
TEST=emerge-samus chromeos-u-boot
CQ-DEPEND=CL:208210,CL:208195

Change-Id: I1334060392b101f5685cff8f9a68a856df80d5d7
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/208196
Reviewed-by: Bill Richardson <wfrichar@chromium.org>