vboot: Add support for new streaming APIs
This implementation is the same as the stub implementations in
vboot_reference, and translates calls to the new streaming APIs to the
old VbExDiskRead() function.  This will allow the vboot_reference
change which uses the new VbExStream*() APIs to commit.

We are not actively developing on u-boot, so we'll likely just leave
this code in u-boot.  It has negligible overhead.

BUG=chromium:403432
BRANCH=none
TEST=The vboot_reference code which calls these APIs hasn't been
     committed yet.  The same code is compiled and tested as part of
     the vboot code unit tests, and boots when compiled into depthcharge,
     so it works in other contexts.
CQ-DEPEND=CL:221992

Change-Id: I43f2d7692a093d13e82747a70c206bb3ea9beb57
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222945
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>