vboot: ec: Check for EC limit power request in VbExEcVbootDone()
Add VbExEcVbootDone(), which checks the limit power flag that gets set by
the EC to request that the AP limit its power use. If this flag remains
set for 3 seconds, request a system shutdown since we're likely to
brownout if we continue to boot.

BUG=chromium:537269
TEST=Compile on daisy
BRANCH=None

Change-Id: Ie03496dfe94410a0026105c076dfa76d1b11aa65
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/309523
Commit-Ready: Shawn N <shawnn@chromium.org>
Tested-by: Shawn N <shawnn@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>